requirements.txt
pandas==2.2.0
openpyxl==3.1.2
matplotlib==3.13.2
PyPDF2==3.0.1
python-docx==1.1.0